fix profile link and private highlight
Have fix the own profile link to yourself. the problem was that since the your name is in the same level of the userlist and other name on the list. so I have also have to install the highlight hover on your own profile to make sure it only get your profile and no one else. I also have finish with the highlighted private chat. it scans all the names in the userlist to see who in private chat with. than it compare their name to the receiver name. If it true, it will highlight the name with darker light blue. After you join private with someone else or turn off private chat. the name should return to normal.(false)